# frontend/specs/001-faq-system-upgrade/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: FAQ System API
  description: |
    DIY Ski FAQ System Backend API - 提供搜尋、Intent 偵測、Slot 擷取、分析和 CRM 整合功能

    **核心功能**:
    - FAQ 搜尋與推薦
    - Intent 偵測（Rule Engine + LLM 增強）
    - Slot 擷取（雪場、日期、人數等）
    - 搜尋分析與統計
    - CRM 整合介面

    **認證**: 目前版本不需要認證（內部使用），未來可加入 API Key
  version: 1.0.0
  contact:
    name: DIY Ski CRM Team
    url: https://diy.ski
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.diy.ski/api/v1
    description: Production server (未來部署)
...
paths:
  /faq/search:
    post:
      tags:
        - FAQ
      summary: 搜尋 FAQ
      description: |
        使用 Fuse.js 進行 FAQ 搜尋，回傳最相關的結果。
        支援多語言查詢（中文、英文、泰文）。

        **效能目標**: <100ms (P95)
      operationId: searchFAQ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  minLength: 1
                  maxLength: 100
                  description: 使用者查詢文字
                  example: "去野澤要準備什麼裝備"
                language:
                  type: string
                  enum: [zh, en, th]
                  default: zh
                  description: 查詢語言
                limit:
                  type: integer
                  minimum: 1
                  maximum: 20
                  default: 5
                  description: 最多回傳幾筆結果
                threshold:
                  type: number
                  minimum: 0
                  maximum: 1
                  default: 0.6
                  description: Fuse.js 分數閾值（0-1，越小越嚴格）
                include_related:
                  type: boolean
                  default: false
                  description: 是否包含相關 FAQ
      responses:
        '200':
          description: 搜尋成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                  - meta
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      results:
                        type: array
                        items:
                          $ref: '#/components/schemas/FAQSearchResult'
                      total:
                        type: integer
                        description: 符合條件的總數
                        example: 5
                  meta:
                    $ref: '#/components/schemas/ResponseMeta'
              example:
                success: true
                data:
                  results:
                    - faq_id: "faq.gear.001"
                      canonical_question: "去滑雪需要準備什麼裝備？"
                      answer_preview: "基本裝備包含雪衣、雪褲、手套..."
                      score: 0.12
                      confidence: 88
                      matched_fields: ["canonical_question", "keywords"]
                  total: 5
                meta:
                  timestamp: "2025-01-15T14:30:00.123Z"
                  query_id: "q_20250115_143000"
                  response_time_ms: 45
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /faq/{faq_id}:
    get:
      tags:
        - FAQ
      summary: 取得單一 FAQ 詳情
      description: 根據 FAQ ID 取得完整內容（含多語言）
      operationId: getFAQById
      parameters:
        - name: faq_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^faq\.[a-z_]+\.\d{3}$'
          example: "faq.gear.001"
        - name: language
          in: query
          schema:
            type: string
            enum: [zh, en, th]
            default: zh
          description: 回傳內容的語言
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/FAQItem'
                  meta:
                    $ref: '#/components/schemas/ResponseMeta'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
    FAQSearchResult:
      type: object
      required:
        - faq_id
        - canonical_question
        - score
        - confidence
      properties:
        faq_id:
          type: string
          example: "faq.gear.001"
        canonical_question:
          type: string
          description: 標準問題（依語言參數回傳對應語言）
          example: "去滑雪需要準備什麼裝備？"
        answer_preview:
          type: string
          description: 答案預覽（前 100 字元）
          example: "基本裝備包含雪衣、雪褲、手套..."
        score:
          type: number
          description: Fuse.js 評分（0-1，越小越好）
          example: 0.12
        confidence:
          type: number
          description: 信心度（0-100）
          example: 88
        matched_fields:
          type: array
          items:
            type: string
          description: 匹配的欄位
          example: ["canonical_question", "keywords"]
        related_faqs:
          type: array
          items:
            type: object
            properties:
              faq_id:
                type: string
              canonical_question:
                type: string
    FAQItem:
      type: object
      required:
        - id
        - intent
        - canonical_question
        - utterance_patterns
        - answer_template
        - keywords
        - crm_tags
        - metadata
      properties:
        id:
          type: string
          pattern: '^faq\.[a-z_]+\.\d{3}$'
          example: "faq.gear.001"
        intent:
          type: string
          enum: [ITINERARY, BOOKING, COURSE, INSTRUCTOR, KIDS_SAFETY, PAYMENT, GEAR, SERVICE, GENERAL]
        canonical_question:
          type: string
          description: 中文標準問題
        canonical_question_en:
          type: string
          description: 英文標準問題
        canonical_question_th:
          type: string
          description: 泰文標準問題
        utterance_patterns:
          type: array
          items:
            type: string
          minItems: 3
        answer_template:
          type: object
          required:
            - text
          properties:
            text:
              type: string
              description: 中文回答（支援 Markdown）
            text_en:
              type: string
            text_th:
              type: string
            postscript:
              type: string
            postscript_en:
              type: string
            postscript_th:
              type: string
            links_inline:
              type: array
              items:
                type: string
        keywords:
          type: array
          items:
            type: string
          minItems: 2
        related_faq_ids:
          type: array
          items:
            type: string
        crm_tags:
          type: array
          items:
            type: string
            pattern: '^#'
        metadata:
          type: object
          required:
            - created_at
            - updated_at
            - version
            - status
          properties:
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            version:
              type: integer
              minimum: 1
            status:
              type: string
              enum: [active, draft, archived]
            priority:
              type: integer
              minimum: 1
              maximum: 10
...
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2025-01-15T14:30:00.123Z"
        query_id:
          type: string
          example: "q_20250115_143000"
        response_time_ms:
          type: integer
          example: 45
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - meta
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: "INVALID_QUERY"
            message:
              type: string
              example: "查詢文字不得為空"
            details:
              type: object
              description: 詳細錯誤資訊（僅開發模式）
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            query_id:
              type: string
  responses:
    BadRequest:
      description: 請求參數錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "INVALID_QUERY"
              message: "查詢文字不得為空"
            meta:
              timestamp: "2025-01-15T14:30:00.123Z"
    NotFound:
      description: 資源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "FAQ_NOT_FOUND"
              message: "找不到指定的 FAQ"
            meta:
              timestamp: "2025-01-15T14:30:00.123Z"
    InternalError:
      description: 伺服器內部錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "INTERNAL_ERROR"
              message: "伺服器發生錯誤，請稍後再試"
            meta:
              timestamp: "2025-01-15T14:30:00.123Z"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key（未來版本啟用）

# 目前版本不需要認證
security: []